Creating Get-CfgSetting function to return a setting using environment inheritance.
Each function had a "Path" parameter to load configuration from a specific file. I redesigned the API so that loading
configuration from a specific file is only done by "Import-CfgConfiguration". The rest of the function will receive a
piped configuration object to operate on, and if they aren't, will use "Import-CfgConfiguration" to load the default
configuration file.

Changed the default configuration file to look for the first baton.json file begining in the current directory
followed by each of its parent directories. This is how most tools work.